+++ web-spike-wasm/src/wasmJsMain/kotlin/dev/syncforge/webspike/WebSpikeMarkers.kt
@@ -1,0 +1,5 @@
+package dev.syncforge.webspike
+
+object WebSpikeMarkers {
+    const val SPIKE_VERSION = "1.6-00"
+}